Tumblr API for posts

+++ app/src/main/java/ru/bulat/mukhutdinov/sample/infrastructure/common/model/BaseModel.kt
@@ -1,3 +1,3 @@
 package ru.bulat.mukhutdinov.sample.infrastructure.common.model
 
-abstract class BaseModel(open val id: Int)
+abstract class BaseModel(open val id: Long)
