+++ app/src/main/java/ru/bulat/mukhutdinov/sample/user/di/UserInjectionModule.kt
@@ -9,10 +9,10 @@ import org.kodein.di.generic.factory
 import org.kodein.di.generic.instance
 import org.kodein.di.generic.singleton
 import ru.bulat.mukhutdinov.sample.infrastructure.common.db.SampleDatabase
-import ru.bulat.mukhutdinov.sample.infrastructure.common.di.ViewModelFactory
+import ru.bulat.mukhutdinov.sample.infrastructure.common.ui.ViewModelFactory
 import ru.bulat.mukhutdinov.sample.user.db.UserDao
-import ru.bulat.mukhutdinov.sample.user.gateway.UserLocalGateway
 import ru.bulat.mukhutdinov.sample.user.gateway.UserDbGateway
+import ru.bulat.mukhutdinov.sample.user.gateway.UserLocalGateway
 import ru.bulat.mukhutdinov.sample.user.ui.UserAndroidViewModel
 import ru.bulat.mukhutdinov.sample.user.ui.UserViewModel
 
@@ -28,14 +28,14 @@ object UserInjectionModule {
             return@singleton UserDbGateway(instance())
         }
 
-        bind<UserViewModel>() with factory { userId: Int, fragment: Fragment ->
+        bind<UserViewModel>() with factory { userId: Long, fragment: Fragment ->
             return@factory ViewModelProviders
                 .of(fragment, UserViewModelFactory(userId, instance()))
                 .get(UserAndroidViewModel::class.java)
         }
     }
 
-    private class UserViewModelFactory(private val userId: Int,
+    private class UserViewModelFactory(private val userId: Long,
                                        private val userLocalGateway: UserLocalGateway) : ViewModelFactory() {
 
         override fun viewModel(): ViewModel =
