+++ app/src/main/java/ru/bulat/mukhutdinov/sample/user/model/User.kt
@@ -2,7 +2,7 @@ package ru.bulat.mukhutdinov.sample.user.model
 
 import ru.bulat.mukhutdinov.sample.infrastructure.common.model.BaseModel
 
-data class User(override val id: Int,
+data class User(override val id: Long,
                 var name: String,
                 var icon: String,
                 var iconThumbnail: String,
